feat:(完善推荐-banner-排行榜日期选择)

+++ ZMBCY/Common/Model/ZMPictureMetadata.h
@@ -2,7 +2,7 @@
 //  ZMPictureMetadata.h
 //  ZMBCY
 //
-//  Created by ZOMAKE on 2017/12/14.
+//  Created by Brance on 2017/12/14.
 //  Copyright © 2017年 Brance. All rights reserved.
 //  图片元数据
 
@@ -13,12 +13,12 @@
 @property (nonatomic, copy) NSString            *imageId;
 /** 宽屏正方形图 */
 @property (nonatomic, copy) NSString            *fullUrl;
-@property (nonatomic, assign) UInt64            width;
+@property (nonatomic, assign) CGFloat            width;
 /* 实际宽度 * 2px */
 @property (nonatomic, assign) CGFloat           realWidth;
-@property (nonatomic, assign) UInt64            height;
+@property (nonatomic, assign) CGFloat            height;
 /* 实际高度 * 2px */
 @property (nonatomic, assign) CGFloat           realHeight;
-@property (nonatomic, assign) UInt64            size;
+@property (nonatomic, assign) CGFloat            size;
 
 @end
